+++ docs/about/release-notes.rst
@@ -16,6 +16,7 @@ Features
 * Added ``unit`` property to ``obj.bins`` for getting and setting unit of bin elements `#2330 <https://github.com/scipp/scipp/pull/2330>`_.
 * Added :py:func:`scipp.optimize.curve_fit` as convenience wrappers of the scipy function of the same name `#2350 <https://github.com/scipp/scipp/pull/2350>`_.
 * Added :py:func:`scipp.signal.butter` and :py:func:`scipp.signal.sosfiltfilt` as wrappers of the scipy functions of the same name `#2356 <https://github.com/scipp/scipp/pull/2356>`_.
+* Added :py:func.`scipp.Variable.to` and :py:func.`scipp.DataArray.to` as a convenience function for simultaneously converting dtype and units.
 
 Breaking changes
 ~~~~~~~~~~~~~~~~